==> codes/00_config_file.R <==
#----------------------------------------------------------#
#
#
#             SAC Diversity and distributions
#
#                   00_config_file.R
#
#                    
#                       2025
#
#----------------------------------------------------------#


#----------------------------------------------------------#
# Variables and paths  -----
#----------------------------------------------------------#

paths <- list(
  input_path = file.path(here(), "data", "input"),
  output_path = file.path(here(), "data", "output"),
  figure_path = file.path(here(), "figures")
)
